requirements.txt
fastmcp>=0.1.0
paho-mqtt>=2.0.0
python-dotenv>=1.0.0
# mem0>=0.2.0  # Commented out as it's not available on PyPI
# psycopg2-binary>=2.9.6  # Only needed for memory server
# openai>=1.0.0  # Only needed for memory server
setuptools>=65.5.1  # Required by some dependencies
flask>=2.0.0  # For the light bulb simulator web server